Make bw_trace_error_handler() return false
After extensive debugging of the problems with WP Engine
I've decided to return false for the time being.
We want the developer to be aware of the message.

The rest of the commented out code is what might be necessary once again
if the bw_trace_obsafe_print_r() solution turns out to be less effective
than expected.

Also note that sometimes the $errcontext structure can be enormous!
So tracing it's output might need to be optional in the future.