Improvements to support different trace levels
The trace level is now included in the trace record.
The previously hardcoded ':0)' after the line number was intended to represent the column
so could be used by some editors to position the cursor.
I've hijacked this to display the trace level used in the call.

bw_lazy_trace2() and bw_lazy_trace() now accept the trace $level parameter

Also changed bw_trace_anonymize_symlinked_file() so it won't issue a Warning
when $real_plugin is null.
Further work needed to find when this problem actually occurs.
But we don't want to break tracing!